+++ packages/orq-rc/src/models/errors/index.ts
@@ -21,7 +21,6 @@ export * from "./honoapierror.js";
 export * from "./httpclienterrors.js";
 export * from "./invokeeval.js";
 export * from "./orqerror.js";
-export * from "./postv2agentsa2a.js";
 export * from "./postv2agentskeycardrefresh.js";
 export * from "./postv2feedback.js";
 export * from "./postv2feedbackremove.js";